feat(frontend): make Private Channel List component
